requirements.txt
pip==22.0.2
# Specifies the version of pip, the package installer for Python, to ensure compatibility with other packages.

cryptography==44.0.2
# A package providing cryptographic recipes and primitives to Python developers. It's used for encryption, decryption, and other security-related tasks.

azure-identity
# A package that provides Azure Active Directory token authentication support, used for authenticating with Azure services.

azure-mgmt-resource
# A package for managing Azure resources, including resource groups, deployments, and templates.

azure-mgmt-network
# A package for managing Azure network resources, such as virtual networks, subnets, and network interfaces.

azure-mgmt-subscription
# A package for managing Azure subscriptions, including listing and retrieving subscription details.